backend/requirements.txt
asgiref==3.9.1
django==5.2.4
graphql-core==3.2.6
packaging==25.0
python-dateutil==2.9.0.post0
six==1.17.0
sqlparse==0.5.3
strawberry-graphql==0.278.0
strawberry-graphql-django==0.65.1
typing-extensions==4.14.1
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.1
pyjwt==2.10.1
django-cors-headers==4.7.0
docker
ipython